taskmanager/pyproject.toml
[project]
name = "taskmanager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "ipykernel>=6.29.5",
    "langchain-core>=0.3.9",
    "langchain-openai>=0.2.1",
    "python-dotenv>=1.0.1",
    "ruff>=0.6.9",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
